Move TimeArgs and init logic to top-level class
